working csl generation
